- fixing routing without local storage. - new component created.
