+++ package.json
@@ -2,7 +2,7 @@
   "name": "frontend-playground",
   "version": "1.0.0",
   "description": "",
-  "main": "main.js",
+  "main": "router.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "webpack serve --open --config webpack.config.js",
