+++ src/views/global-search.html
@@ -1,0 +1,15 @@
+<template>
+    <main class="home row justify-content-center">
+       <h4>amabatukaaaaaaam</h4>
+    </main>
+</template>
+<script>
+    (function components(){
+            console.log('fokiu mai bro');
+    })()
+</script>
+<style>
+    h4 {
+        color: red;
+    }
+</style>
